Initial commit of the revamped GPDB utilities
This commit represents the first stage of a broad effort to rewrite the
existing variety of GPDB Python utilities into a single unified utility
in Go, to which functionality will be added over time as more utilities
are converted to the new framework.

This new "gp" binary works on a hub-agent model, in which tasks are
executed by a hub process running on the coordinator host and a set of
agent processes running on each segment host.  Where the current
utilities carry out work on the segments via SSH, this framework uses
gRPC to communicate between the hub and agents and between the hub and
the gp binary itself, removing passwordless SSH access to the segment
hosts as a requirement for GPDB cluster operation.

Where the current utilities are scripts that need to do their own setup
and teardown each time they are invoked, the hub and agents are
long-running processes orchestrated through systemd (on Linux) and
launchd (on OSX) that will be used by every utility call.

Currently, gp only supports the ability to start and stop the hub and
agent processes and poll their status, so that they can serve as a proof
of concept for the new framework before other utilities are ported to
it.  These utilities will be converted one at a time, beginning with
gpinitsystem, and will be added to gp as top-level CLI commands.

This commit does *not* add the necessary Golang dependencies to GPDB,
hook this Makefile into the main GPDB build process, or add jobs for gp
tests to the main GPDB pipeline.  That will be done once this framework
is considered beta-ready.

Co-authored-by: Jamie McAtamney <jmcatamney@vmware.com>
Co-authored-by: Nihal Jain <jnihal@vmware.com>
Co-authored-by: Piyush Chandwadkar <pchandwadkar@vmware.com>
Co-authored-by: Rakesh Sharma <shrakesh@vmware.com>