gpcheckperf: add buffer size parameter (#15861)
Before this patch, while running gpcheckperf utility the buffer was set by
default for the underlying gpnetbenchClient utility as 32Kb. It led to problem
with receiving annoying and misleading warnings about connections between
hosts

Use '--buffer-size' flag with size in kilobytes to set buffer size,
which will be used at gpnetbenchClient.
It is an optional parameter. The default value is 32Kb.