Change `segment-array` structure
This commit changes the way we take the primary segments from the user. Instead of only accepting an array of primary segments, we will accept an array of segment pairs and each pair will be a structure containing its corresponding primary and mirror segments. This will make it easier to take inputs when we support mirrored configuration as well.