Fixing if the time command has comma in the output
this is to fix gpcheckperf fails with an exception when sometime time commmand output
has a comma separator instead of a dot.
Steps to reproduce issue:
$export LC_ALL=de_DE_utf8
$time sleep 1
real	0m1,021s
user	0m0,001s
sys	0m0,005s

Fix: added check if comma present in the time output, replace it with a dot and continue parsing.

Testing: Added unit tests to check the output of parseMultiDDResult() in case of comma and dot.